revision

+++ vector.h
@@ -3,8 +3,8 @@
 class Vector2 {
 public:
 	Vector2();
-	Vector2(int x, int y);
-	int x, y;
+	Vector2(float x, float y);
+	float x, y;
 	static const Vector2 Zero;
 	Vector2& operator+=(const Vector2& right)
 	{
